+++ CHANGELOG.md
@@ -7,12 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 ### Added
 - `Year`, `Month`, `Date`, `Time`, `LocalDateTime` and `ZonedDateTime` are comparable from now on.
-  Next methods had added to these types:
+  Next methods were added to these types:
     - `compare()`
     - `greaterThan()`
     - `greaterThanOrEqual()`
     - `lessThan()`
     - `lessThanOrEqual()`
+- Value objects were made serializable: `Year`, `Date`, `Time`, `LocalDateTime`, `ZonedDateTime`, `TimeZone`,
+  `TimeZoneId`, `Offset`.
 
 ## v0.1.1 - 2016-10-10
 ### Added